Add files via upload
Hello! This is an example of a validation form, where the customer can send a message to the owner of the website. I wrote it using HTML and PHP but if you want to improve the texture of the form just go to my HTML/CSS repository and you will find there several examples of forms, according to everyone's tastes. Have a nice day!